+++ src/main/java/be/ugent/rml/MappingFactory.java
@@ -35,10 +35,21 @@ public class MappingFactory {
     // This boolean is true when the double in a reference need to be ignored.
     // For example, when accessing data in a RDB.
     private boolean ignoreDoubleQuotes;
+
+    // Base IRI to prepend to a relative IRI to make it absolute.
+    private final String baseIRI;
+
+    // StrictMode determines RMLMapper's behaviour when an IRI for a NamedNode is invalid.
+    // If set to BEST_EFFORT, RMLMapper will not generate a NamedNode and go on.
+    // If set to STRICT, RMLMapper will stop execution with an exception.
+    private final StrictMode strictMode;
+
     protected Logger logger = LoggerFactory.getLogger(this.getClass());
 
-    public MappingFactory(Agent functionAgent) {
+    public MappingFactory(final Agent functionAgent, final String baseIRI, final StrictMode strictMode) {
         this.functionAgent = functionAgent;
+        this.baseIRI = baseIRI;
+        this.strictMode = strictMode;
     }
 
     public Mapping createMapping(Term triplesMap, QuadStore store) throws Exception {
@@ -90,7 +101,7 @@ public class MappingFactory {
                         }
                     } else {
                         //we are not dealing with a Blank Node, so we create the template
-                        generator = new NamedNodeGenerator(RecordFunctionExecutorFactory.generate(store, subjectmap, true, ignoreDoubleQuotes));
+                        generator = new NamedNodeGenerator(RecordFunctionExecutorFactory.generate(store, subjectmap, true, ignoreDoubleQuotes), baseIRI, strictMode);
                     }
                 } else {
                     SingleRecordFunctionExecutor functionExecutor = parseFunctionTermMap(functionValues.get(0));
@@ -98,7 +109,7 @@ public class MappingFactory {
                     if (isBlankNode) {
                         generator = new BlankNodeGenerator(functionExecutor);
                     } else {
-                        generator = new NamedNodeGenerator(functionExecutor);
+                        generator = new NamedNodeGenerator(functionExecutor, baseIRI, strictMode);
                     }
                 }
 
@@ -116,8 +127,8 @@ public class MappingFactory {
                      * Don't put in graph for rr:class, subject is already put in graph, otherwise double export.
                      * Same holds for targets, the rdf:type triple will be exported to the subject target already.
                      */
-                    NamedNodeGenerator predicateGenerator = new NamedNodeGenerator(new ConstantExtractor(NAMESPACES.RDF + "type"));
-                    NamedNodeGenerator objectGenerator = new NamedNodeGenerator(new ConstantExtractor(c.getValue()));
+                    NamedNodeGenerator predicateGenerator = new NamedNodeGenerator(new ConstantExtractor(NAMESPACES.RDF + "type"), baseIRI, strictMode);
+                    NamedNodeGenerator objectGenerator = new NamedNodeGenerator(new ConstantExtractor(c.getValue()), baseIRI, strictMode);
                     predicateObjectGraphMappings.add(new PredicateObjectGraphMapping(
                             new MappingInfo(subjectmap, predicateGenerator),
                             new MappingInfo(subjectmap, objectGenerator),
@@ -187,7 +198,7 @@ public class MappingFactory {
             if (o instanceof Literal) {
                 gen = new LiteralGenerator(fn);
             } else {
-                gen = new NamedNodeGenerator(fn);
+                gen = new NamedNodeGenerator(fn, baseIRI, strictMode);
             }
 
             // rr:object shortcut can never have targets
@@ -223,7 +234,7 @@ public class MappingFactory {
                         oGen = new LiteralGenerator(executor);
                     }
                 } else if (termType.equals(new NamedNode(NAMESPACES.RR + "IRI"))) {
-                    oGen = new NamedNodeGenerator(executor);
+                    oGen = new NamedNodeGenerator(executor, baseIRI, strictMode);
                 } else {
                     if (executor == null) {
                         // This will generate Blank Node with random identifiers.
@@ -327,7 +338,7 @@ public class MappingFactory {
                     gen = new LiteralGenerator(functionExecutor);
                 }
             } else {
-                gen = new NamedNodeGenerator(functionExecutor);
+                gen = new NamedNodeGenerator(functionExecutor, baseIRI, strictMode);
             }
 
             // get targets for object map
@@ -361,7 +372,7 @@ public class MappingFactory {
                 SingleRecordFunctionExecutor executor = RecordFunctionExecutorFactory.generate(store, graphMap, true, ignoreDoubleQuotes);
 
                 if (termType == null || termType.equals(new NamedNode(NAMESPACES.RR + "IRI"))) {
-                    generator = new NamedNodeGenerator(executor);
+                    generator = new NamedNodeGenerator(executor, baseIRI, strictMode);
                 } else {
                     if (executor == null) {
                         generator = new BlankNodeGenerator();
@@ -373,7 +384,7 @@ public class MappingFactory {
                 SingleRecordFunctionExecutor functionExecutor = parseFunctionTermMap(functionValues.get(0));
 
                 if (termType == null || termType.equals(new NamedNode(NAMESPACES.RR + "IRI"))) {
-                    generator = new NamedNodeGenerator(functionExecutor);
+                    generator = new NamedNodeGenerator(functionExecutor, baseIRI, strictMode);
                 } else {
                     generator = new BlankNodeGenerator(functionExecutor);
                 }
@@ -390,7 +401,7 @@ public class MappingFactory {
         for (Term graph : graphShortcuts) {
             String gStr = graph.getValue();
             // rr:graph shortcut can never have targets
-            graphMappingInfos.add(new MappingInfo(termMap, new NamedNodeGenerator(new ConstantExtractor(gStr))));
+            graphMappingInfos.add(new MappingInfo(termMap, new NamedNodeGenerator(new ConstantExtractor(gStr), baseIRI, strictMode)));
         }
 
         return graphMappingInfos;
@@ -410,12 +421,12 @@ public class MappingFactory {
 
             if (functionValues.isEmpty()) {
                 predicateMappingInfos.add(new MappingInfo(predicateMap,
-                        new NamedNodeGenerator(RecordFunctionExecutorFactory.generate(store, predicateMap, false, ignoreDoubleQuotes)),
+                        new NamedNodeGenerator(RecordFunctionExecutorFactory.generate(store, predicateMap, false, ignoreDoubleQuotes), baseIRI, strictMode),
                         targets));
             } else {
                 SingleRecordFunctionExecutor functionExecutor = parseFunctionTermMap(functionValues.get(0));
 
-                predicateMappingInfos.add(new MappingInfo(predicateMap, new NamedNodeGenerator(functionExecutor), targets));
+                predicateMappingInfos.add(new MappingInfo(predicateMap, new NamedNodeGenerator(functionExecutor, baseIRI, strictMode), targets));
             }
         }
 
@@ -424,7 +435,7 @@ public class MappingFactory {
         for (Term predicate : predicateShortcuts) {
             String pStr = predicate.getValue();
             // rr:predicate shortcut can never have targets
-            predicateMappingInfos.add(new MappingInfo(termMap, new NamedNodeGenerator(new ConstantExtractor(pStr))));
+            predicateMappingInfos.add(new MappingInfo(termMap, new NamedNodeGenerator(new ConstantExtractor(pStr), baseIRI, strictMode)));
         }
 
         return predicateMappingInfos;